refactor(core): Split commands/index.ts into domain-based factory modules for LLM optimization
Split the 890-line commands/index.ts into focused factory modules:
- factories/types.ts - CommandFactory type definition
- factories/movement.ts - Move, Explore factories
- factories/cards.ts - PlayCard, PlayCardSideways, ResolveChoice factories
- factories/combat.ts - Combat-related factories (EnterCombat, EndCombatPhase, etc.)
- factories/units.ts - RecruitUnit, ActivateUnit factories
- factories/turn.ts - EndTurn, Rest, AnnounceEndOfRound factories
- factories/sites.ts - Interact, EnterSite, ResolveGladeWound, ResolveDeepMine factories
- factories/tactics.ts - Tactic-related factories
- factories/offers.ts - BuySpell, LearnAdvancedAction, SelectReward factories
- factories/index.ts - Registry and re-exports

The main index.ts is now a thin file (~210 lines) that imports the registry
and provides createCommandForAction plus re-exports.
